More code cleaning, SDL window can now scale by integer amounts

+++ util.c
@@ -68,8 +68,8 @@ void time_print(struct timeval *tv)
     char buffer[30];
     time_t curtime;
 
-    printf("%ld.%06ld", tv->tv_sec, tv->tv_usec);
+    printf("%d.%06d", (int)tv->tv_sec, (int)tv->tv_usec);
     curtime = tv->tv_sec;
     strftime(buffer, 30, "%m-%d-%Y  %T", localtime(&curtime));
-    printf(" = %s.%06ld\n", buffer, tv->tv_usec);
+    printf(" = %s.%06d\n", buffer, (int)tv->tv_usec);
 }
